Rol de Supervisor 2
Los cambios que se hicieron fueron para que el rol de Supervisor 2 pueda ser capaz solo de acceder a sus vistas, y ninguna otras más.

+++ resources/views/encuestadores/table.blade.php
@@ -11,7 +11,9 @@
       
                         <!-- Botones de la parte superior derecha -->
                         <div class="box-tools pull-right">
-                            {!! Form::open(['route' => ['encuestadores.destroy', $encuestador->id], 'method' => 'delete']) !!}
+                            @if(Auth::user()->hasRole('Super Admin', 'Supervisor 1'))
+                                {!! Form::open(['route' => ['encuestadores.destroy', $encuestador->id], 'method' => 'delete']) !!}
+                            @endif    
                                 <div class='btn-group'>
 
                                     <!-- Boton para ver los datos del encuestador -->
@@ -25,9 +27,10 @@
                                             <i class="glyphicon glyphicon-edit"></i>
                                         </a>
                     
-                                        <!-- Boton para eliminar los datos del encuestador -->
-                                        {!! Form::button('<i class="glyphicon glyphicon-trash"></i>', ['type' => 'submit', 'class' => 'btn btn-danger btn-xs','onclick' => "return confirm('¿Está seguro de querer eliminar?')", 'data-toggle' => 'modal', 'data-target' => '#modal-danger']) !!}
-
+                                        @if(Auth::user()->hasRole('Super Admin', 'Supervisor 1'))
+                                            <!-- Boton para eliminar los datos del encuestador -->
+                                            {!! Form::button('<i class="glyphicon glyphicon-trash"></i>', ['type' => 'submit', 'class' => 'btn btn-danger btn-xs','onclick' => "return confirm('¿Está seguro de querer eliminar?')", 'data-toggle' => 'modal', 'data-target' => '#modal-danger']) !!}
+                                        @endif
                                     {{-- @endif --}}
 
                                     <!-- Boton para minimizar/maximiar cada cuadro -->
@@ -36,7 +39,9 @@
                                     </button>
       
                                 </div>
-                            {!! Form::close() !!}
+                            @if(Auth::user()->hasRole('Super Admin', 'Supervisor 1'))
+                                {!! Form::close() !!}
+                            @endif
                         </div>
                     </div>
